Compile all files for diagnostics
This fixes an issue where diagnostics from only the open files would
show up, by doing pretty much the same as elm-tooling's language server
is doing.
https://github.com/elm-tooling/elm-language-server/pull/629

The approach I've used here is simpler though, as it doesn't care to
distinguish between test and regular files, so if test dependencies are
imported in non-test files, the imports are not reported as errors.
Imo this is totally fine, as doing this by mistake is rare in my
experience and even if that's done, elm-watch or elm make or whatever is
used will catch the error and it's easy to fix.